% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLclhostEng.R
\name{setLclhostEng}
\alias{setLclhostEng}
\title{Set noproxy localhost and environment to english}
\usage{
setLclhostEng(http_msg = FALSE)
}
\arguments{
\item{http_msg}{Want to see localhost http msgs? Defaults to FALSE [don't show msgs].}
}
\description{
This function sets localhost to noproxy and sys enviroment to english.
}
\examples{
setLclhostEng(http_msg = FALSE)
setLclhostEng(http_msg = TRUE)
setLclhostEng()
}
\keyword{localhost_english}

